split tests, move instance eval to top level
