+++ lib/massive_sitemap/version.rb
@@ -1,3 +1,3 @@
 module MassiveSitemap
-  VERSION = "2.0.1"
+  VERSION = "2.0.0"
 end
